+++ plugin-build/plugin/src/test/java/com/ncorti/ktfmt/gradle/tasks/KtfmtFormatTaskIntegrationTest.kt
@@ -36,6 +36,9 @@ internal class KtfmtFormatTaskIntegrationTest {
 
         assertThat(result.task(":ktfmtFormatMain")?.outcome).isEqualTo(FAILED)
         assertThat(result.output).containsMatch("Failed to format file: .*TestFile.kt \\(reason =")
+        assertThat(result.output).contains("Ktfmt failed to run with 1 failures")
+        assertThat(result.output)
+            .contains("src${File.separator}main${File.separator}java${File.separator}TestFile.kt")
     }
 
     @Test